Split assoc items out of `trait_data`/`impl_data` queries
