+++ crates/hir-ty/src/display.rs
@@ -1342,7 +1342,7 @@ impl HirDisplay for Ty {
                             .lang_item(body.module(db.upcast()).krate(), LangItem::Future)
                             .and_then(LangItemTarget::as_trait);
                         let output = future_trait.and_then(|t| {
-                            db.trait_data(t).associated_type_by_name(&Name::new_symbol_root(
+                            db.trait_items(t).associated_type_by_name(&Name::new_symbol_root(
                                 sym::Output.clone(),
                             ))
                         });
