+++ crates/hir-ty/src/mir/lower/as_place.rs
@@ -195,7 +195,7 @@ impl MirLowerCtx<'_> {
                                     self.resolve_lang_item(LangItem::DerefMut)?.as_trait()
                                 {
                                     if let Some(deref_fn) =
-                                        self.db.trait_data(deref_trait).method_by_name(
+                                        self.db.trait_items(deref_trait).method_by_name(
                                             &Name::new_symbol_root(sym::deref_mut.clone()),
                                         )
                                     {
@@ -353,7 +353,7 @@ impl MirLowerCtx<'_> {
             .ok_or(MirLowerError::LangItemNotFound(trait_lang_item))?;
         let deref_fn = self
             .db
-            .trait_data(deref_trait)
+            .trait_items(deref_trait)
             .method_by_name(&trait_method_name)
             .ok_or(MirLowerError::LangItemNotFound(trait_lang_item))?;
         let deref_fn_op = Operand::const_zst(
